pyproject.toml
[tool.poetry]
name = "mwa-chala"
version = "0.1.0"
description = "portfolio website"
authors = ["mwachala <berrybretch@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0.1"
httpserver = "^1.1.0"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
